+++ packages/ui-mobile-visreg/src/routes.ts
@@ -400,6 +400,11 @@ export const routes = [
       require('@coinbase/cds-mobile/overlays/__stories__/ModalBasic.stories').default,
   },
   {
+    key: 'ModalCustomPadding',
+    getComponent: () =>
+      require('@coinbase/cds-mobile/overlays/__stories__/ModalCustomPadding.stories').default,
+  },
+  {
     key: 'ModalLong',
     getComponent: () =>
       require('@coinbase/cds-mobile/overlays/__stories__/ModalLong.stories').default,