+++ packages/web/src/overlays/modal/ModalHeader.tsx
@@ -1,18 +1,19 @@
 import React from 'react';
 import { useModalContext } from '@coinbase/cds-common/overlays/ModalContext';
 import { interactableHeight } from '@coinbase/cds-common/tokens/interactableHeight';
-import type {
-  SharedAccessibilityProps,
-  SharedProps,
-  ValidateProps,
-} from '@coinbase/cds-common/types';
+import type { SharedAccessibilityProps } from '@coinbase/cds-common/types';
 
 import { IconButton } from '../../buttons/IconButton';
-import { Box, type BoxDefaultElement, type BoxProps } from '../../layout/Box';
-import { HStack } from '../../layout/HStack';
+import { Box } from '../../layout/Box';
+import {
+  HStack,
+  type HStackBaseProps,
+  type HStackDefaultElement,
+  type HStackProps,
+} from '../../layout/HStack';
 import { Text } from '../../typography/Text';
 
-export type ModalHeaderBaseProps = {
+export type ModalHeaderBaseProps = Omit<HStackBaseProps, 'children'> & {
   /** Handles back button press */
   onBackButtonClick?: React.MouseEventHandler;
   /** Title of the Modal */
@@ -51,9 +52,10 @@ export type ModalHeaderBaseProps = {
    * @link https://reactnative.dev/docs/accessibility#accessibilityhint
    */
   closeAccessibilityHint?: SharedAccessibilityProps['accessibilityHint'];
-} & SharedProps;
+};
 
-export type ModalHeaderProps = ModalHeaderBaseProps & BoxProps<BoxDefaultElement>;
+export type ModalHeaderProps = ModalHeaderBaseProps &
+  Omit<HStackProps<HStackDefaultElement>, 'children'>;
 
 export const ModalHeader = ({
   alignItems = 'center',
@@ -83,10 +85,7 @@ export const ModalHeader = ({
       borderedBottom={!hideDividers}
       paddingX={paddingX}
       paddingY={paddingY}
-      {...(props satisfies ValidateProps<
-        typeof props,
-        Omit<ModalHeaderProps, keyof BoxProps<BoxDefaultElement>>
-      >)}
+      {...props}
     >
       {onBackButtonClick ? (
         <Box>