+++ packages/common/CHANGELOG.md
@@ -8,7 +8,9 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
-## Unreleased
+## 8.56.0 ((3/23/2026, 06:31 AM PST))
+
+This is an artificial version bump with no new change.
 
 #### 📘 Misc