+++ packages/mcp-server/CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 8.56.0 ((3/23/2026, 06:31 AM PST))
+
+This is an artificial version bump with no new change.
+
 ## 8.55.1 ((3/22/2026, 01:43 PM PST))
 
 This is an artificial version bump with no new change.